# Performance drop when object size is large