--- a/worker/workItems.go
+++ b/worker/workItems.go
@@ -135,7 +135,7 @@ func (op Stopper) Prepare() error {
 func (op ReadOperation) Do() error {
 	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).WithField("Preexisting?", op.WorksOnPreexistingObject).Debug("Doing ReadOperation")
 	start := time.Now()
-	err := getObject(svc, op.ObjectName, op.Bucket)
+	err := getObject(svc, op.ObjectName, op.Bucket, op.ObjectSize)
 	duration := time.Since(start)
 	promLatency.WithLabelValues(op.TestName, "GET").Observe(float64(duration.Milliseconds()))
 	if err != nil {