… examples/Bandwidth.png
… examples/Bendwidth.png
… examples/Latency.png