Because the allocation of memory before getobject, the performance of read decreases when the object size is large. 

**before**
  | write |   | read |  
-- | -- | -- | -- | --
Object Size | ResTime(Avg) | Bandwidth(MB/S) | ResTime(Avg) | Bandwidth(MB/S)
4KB | 2.756562 | 19.0927658 | 2.303125 | 21.4533014
16KB | 3.016875 | 71.224057 | 1.61375 | 115.489441
64KB | 3.85125 | 229.491497 | 1.996563 | 395.550388
256KB | 7.869375 | 476.363095 | 3.440937 | 1016.75673
1M | 29.525938 | 531.736959 | 15.9925 | 965.220499
4MB | 108.144688 | 587.670134 | 283.579687 | 224.96259

When the size of object is larger than 1MB, the bandwidth of read decreases. It is because of the overhead of memory allocation. 

**after**

  | write |   | read |  
-- | -- | -- | -- | --
Object Size | ResTime(Avg) | Bandwidth(MB/S) | ResTime(Avg) | Bandwidth(MB/S)
4KB | 2.99 | 17.91 | 2.17 | 21.3
16KB | 3.18 | 68.13 | 1.25 | 140.33
64KB | 4.07 | 218.05 | 1.4 | 515.26
256KB | 7.88 | 474.95 | 2.28 | 1424.34
1M | 28.66 | 546.41 | 6.83 | 2168.6
4MB | 108.5 | 584.53 | 31.13 | 2011.05

After changing memory allocation, we see the improvement with latency and bandwidth. 
and I add the comparison of COSBench and GOSBench in README. I hope it is merged.