@mulbc 2022-05-11
Sorry for taking so long - I must have overlooked the notification. 
Do you remember why you added the worker binary to the repo? The rest looks good to me